requirements.txt
aiohttp==3.6.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
crypto==1.4.1
cryptography==2.5
cycler==0.10.0
decorator==4.4.0
docopt==0.6.2
Flask==1.0.2
Flask-Cors==3.0.7
graphviz==0.10.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==1.0.1
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
multidict==4.7.5
Naked==0.1.31
networkx==2.3
nltk==3.6.2
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.2
Pillow==6.0.0
pipreqs==0.4.9
polygon-api-client==0.1.9
praw==7.2.0
prawcore==2.0.0
pycparser==2.19
pyparsing==2.3.1
PySocks==1.6.8
python-dateutil==2.8.0
python-engineio==3.12.1
python-socketio==4.5.0
pytz==2018.9
PyWavelets==1.0.3
PyYAML==5.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.2.0
schedule==1.0.0
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
selenium==3.141.0
shellescape==3.4.1
six==1.12.0
sklearn==0.0
splinter==0.14.0
tqdm==4.60.0
tweepy==3.7.0
update-checker==0.18.0
urllib3==1.24.1
vaderSentiment==3.2.1
websocket-client==0.58.0
websockets==8.1
Werkzeug==0.15.2
xdot==1.1
yarg==0.1.9
yarl==1.4.2
